Changes to Equation and added button that spawns in equation object
I wanted to parse the equation as both the left side and the right side so they player could drag in both side. But i am now using a button to spawn the equation onto the whiteboard. After dealing with spawning issues and collision issues this is the more streamlined way